requirements.txt
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.2
django-heroku==0.3.1
gunicorn==19.7.1
idna==2.6
psycopg2==2.7.3.2
PyGithub==1.35
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
requests==2.18.4
urllib3==1.22
whitenoise==4.1.2